requirements.txt
#PyPDF2==3.0.1
pymupdf==1.25.5
pandas==2.1.4
sentence-transformers==4.0.2
huggingface_hub==0.30.1
accelerate==1.6.0
peft==0.15.1
bitsandbytes
transformers==4.50.3
FlagEmbedding==1.3.4
streamlit==1.30.0
openai==1.70.0
json_repair==0.40.0
hopsworks[python]
python-dotenv
